Fix deploy script and deploy article

+++ deploy.sh
@@ -1,7 +1,12 @@
 #!/bin/sh
 
 USER=dementor
-HOST=azkaban.in
+if [[ -z "$SOUMYADEEP_IP" ]]; then
+    echo "SOUMYADEEP_IP is set to the empty string"
+    exit
+fi
+
+HOST=$SOUMYADEEP_IP
 DIR=/home/dementor/pensieve/
 
 hugo && rsync -avz --delete public/ ${USER}@${HOST}:${DIR}
